// src/app/core/models/unidad-muestreo.ts
import { Falla } from "./falla";

export class UnidadMuestreo {
  idUnidadMuestreo: number | undefined;
  indexUnidad: number | undefined;
  idMantenimientoVialEvento: number | undefined;
  abscisaInicial: number | undefined;
  abscisaFinal: number | undefined;
  area: number | undefined;
  ancho: number | undefined;
  pci: number | undefined;
/*
  tabFallas: Falla[];

  constructor (index: number | undefined, idSolicitudSeleccionada: number | undefined) {
    this.indiceUnidad = index;
    this.tabMantenimientoVial = idSolicitudSeleccionada;
    this.tabFallas = new Array<Falla>();
  }
*/
}